+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add query `_search` for searching entries by text using a search adapter passed to `createClient()`
+- Add directive `@Searchable` to explicitly mark types and fields as searchable
+
 ## [2.0.0-beta.1] - 2026-09-24
 
 ### Changed